MusicSpiralRepresentation

Documentation for MusicSpiralRepresentation.

Spiral Representation

Center of Effect

Estimating the most likely key

MusicSpiralRepresentation.funhar_seq

MusicSpiralRepresentation.get_Major_chord

MusicSpiralRepresentation.get_Major_key

MusicSpiralRepresentation.get_center_effect

MusicSpiralRepresentation.get_cfpitch

MusicSpiralRepresentation.get_cfpitch_mod12

MusicSpiralRepresentation.get_local_lin_w

MusicSpiralRepresentation.get_minor_chord

MusicSpiralRepresentation.get_minor_key

MusicSpiralRepresentation.get_onon_notes

MusicSpiralRepresentation.get_piece_by_measure_csv

MusicSpiralRepresentation.get_piece_by_measure_m21

MusicSpiralRepresentation.get_pitch

MusicSpiralRepresentation.get_rank_freq

MusicSpiralRepresentation.funhar_seq — Method

funhar_seq(kseq, fun_key)
Returns a sequence of Roman numerals given a reference for the Tonic Key.

MusicSpiralRepresentation.get_Major_chord — Method

get_Major_chord(k; w=[0.536, 0.274, 0.19], r=1, h=sqrt(2/15))

Returns the location (x,y,z) of a Major chord given by a linear combination
of a triad constructed from a fundamental note "k", the fifth (k+1) and the third (k+4)
in the spiral representation. The vector "w" contains the weights used for each term.

MusicSpiralRepresentation.get_Major_key — Method

get_Major_key(k; o=[0.536, 0.274, 0.19], w=[0.536, 0.274, 0.19], r=1, h=sqrt(2/15))

Returns the location (x,y,z) of a Major Key given by a linear combination
of chord triads constructed from the major chords of a fundamental "tonic" note "k",
the dominant/fifth (k+1) and the subdominant/minor third (k-3) in the spiral representation.
The vector "o" contains the weights used for each chord and "w" are the weights on each note in the chord.

MusicSpiralRepresentation.get_center_effect — Method

get_center_effect(chunk_notes; r=1, h=sqrt(2/15), 
                all_keys=all_keys, pos_all_keys=pos_all_keys, 
                sbeat_w=[[1.],[1.]], lin_w=1)

Computes the coordinates (x,y,z) for the center of effect (mass)
for a given set of notes, these notes can be in different formats:
Notes are always represented in MIDI notation, note durations can
be in any numeric format.

    Matrix: From the DataFrame built with get_xml_df or get_csv_df
    1d - Array: [n1,n2,...,nm], where each element represents a note in MIDI notation
    two 1d - Array: [n1,n2,...,nm], [d1,d2,...,dm], with the set of n
    representing the notes in MIDI notation and d each of their durations.

MusicSpiralRepresentation.get_cfpitch — Method

get_cfpitch(pitch_seq)

Converts a pitch sequence of MIDI values (0-127) into a pitch sequence ordered by fifths
to be consistent with the cylindrical representation, starting in C=0, G=1, ..., etc.

MusicSpiralRepresentation.get_cfpitch_mod12 — Method

get_cfpitch_mod12(pitch_seq)

Converts a pitch sequence of MIDI values (0-127) into a pitch sequence in mod12 ordered by fifths
to be consistent with the cylindrical representation, starting in C=0, G=1, ..., etc.

MusicSpiralRepresentation.get_local_lin_w — Method

get_local_lin_w(ptcs, factor)

Returns a list of pitches and their respective weights from a sequence of pitches (ptcs)
and a weighting factor (factor), the weighting gives more importance to lower notes
within the 1-factor range.

MusicSpiralRepresentation.get_minor_chord — Method

get_minor_chord(k; u=[0.536, 0.274, 0.19], r=1, h=sqrt(2/15))

Returns the location (x,y,z) of a Minor chord given by a linear combination
of a triad constructed from a fundamental note "k", the fifth (k+1) and the minor third (k-3)
in the spiral representation. The vector "u" contains the weights used for each term.

MusicSpiralRepresentation.get_minor_key — Method

get_minor_key(k; v=[0.536, 0.274, 0.19], w=[0.536, 0.274, 0.19],
u=[0.536, 0.274, 0.19], a=0.75, b=0.75, r=1, h=sqrt(2/15))

Returns the location (x,y,z) of a Minor Key given by a linear combination
of chord triads constructed from the major chords of a fundamental note "k",
the possible "dominant" triad (k+1) and the possible "subdominant" triad (k-3) in the spiral representation.
The vector "v" contains the weights used for each chord and "w" are the weights on each note in the major chords.
"u" are the weights for the minor chords. The dominant and subdominant minor/major chords are weighted by "a" and "b".

MusicSpiralRepresentation.get_onon_notes — Method

get_onon_notes(s)

MusicSpiralRepresentation.get_piece_by_measure_csv — Method

get_piece_by_measure_csv(s; qdiv=32, abs_time=true)

Returns a vector of two dimensional arrays of the basic information of the piece for analyzing its properties,
input "s" should be a CSV table from a CSV file converted from a MIDI file with the midicsv script available for free
in: https://www.fourmilab.ch/webtools/midicsv/ .

The first output consist of a vector which components represent each measure of the music piece,
each measure contains four columns; Beat: number of beat where the note starts (beat or halfbeat)
Measure: the fraction of the measure (n/d), Duration: the duration of the note in midiclocks, if abs_time=true
the output returns the time in ms where the note started and ended before the duration of the note,
Pitch: the pitch of the note represented in MIDI notation (0-127).

The second output corresponds to the total number of notes in the piece and the number of notes that
fall outside the threshold given (1/qdiv), most of the time if notes fall outside this threshold is because
the MIDI was sequenced (recorded) instead of generated with a music score software.

MusicSpiralRepresentation.get_piece_by_measure_m21 — Method

function get_piece_by_measure_m21(m21_data; start_measure=1)

Returns an array with the information of pitches and duration on each measure in the following format:

| #measure | time signature | start_quarter | end_quarter | duration (in quarters) | pitch (0-127) |

MusicSpiralRepresentation.get_pitch — Method

get_pitch(k; r=1, h=sqrt(2/15))

Returns the location (x,y,z) of a pitch "k"
in the spiral representation, pitches are in module 12
the variable "r" is the radius of the cylinder
and "h" is the vertical distance between thirds.

MusicSpiralRepresentation.get_rank_freq — Method

get_rank_freq(series)

Returns a 2-Dimensional Array of symbols and their respective frequencies
from a series of data. The output is ordered by rank (most to least frequent).